Fix Humidity to Decimal & deleted .mdf reference
Fix Humidity to Decimal and deleted mdf refrence

+++ WeatherApp/WeatherData.cs
@@ -34,7 +34,7 @@ namespace WeatherApp
         {
             public double temp { get; set; }
             public double pressure { get; set; }
-            public int humidity { get; set; }
+            public double humidity { get; set; }
             public double temp_min { get; set; }
             public double temp_max { get; set; }
         }
